+++ Database/WhitePage.MainDb/Ops/Tables/trCaseOffender.sql
@@ -3,7 +3,7 @@
 	[CaseOffenderId] INT NOT NULL PRIMARY KEY IDENTITY (1,1),
 	CaseId INT NOT NULL,
 	Name varchar(200),
-	Age smallint,
+	Age DECIMAL(5, 2),
 	GenderLookupId INT,
 	RelationshipWithVictimLookupId INT, 
     OtherRelationship varchar(200)
