+++ Database/WhitePage.MainDb/Ops/User Defined Types/CaseOffenderType.sql
@@ -3,7 +3,7 @@
 	[CaseOffenderId] INT,
 	CaseId INT,
 	Name varchar(200),
-	Age smallint,
+	Age DECIMAL(5, 2),
 	GenderLookupId INT,
 	RelationshipWithVictimLookupId INT,
 	OtherRelationship varchar(200)
